requirements.txt
bing_image_downloader==1.1.2
keras==2.13.1
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.3.0
plotly==5.9.0
scikit_learn==1.2.2
streamlit==1.31.1
streamlit_antd_components==0.3.2
streamlit_shadcn_ui==0.1.18
